🚚 Rename: 하단 탭바 파일 이름 변경
