+++ src/navigation/index.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
 import SignupRoute from './route/signup';
-import BottomTabBar from './tabs/BottomTabBar';
+import BottomTabBar from './tabs';
 
 import LoginScreen from '@/screens/login';
 import OnboardingScreen from '@/screens/onboarding';
